Added IntergerOf
IntegerOf as a class is how we wrap primitive data values into an object. This allows us to use objects in our code and not pass around data structures.
We are doing object oriented programming, not data structure programming. We remove the data storage aspect by encapsulating the data into a class with controlable behavior.

Other reasons to not use the data structure - God Class. Does MANY things.