Migrate test to use IntegerOf
We can first migrate our test code to use Interger and IntegerOf.
With the implicit cast, we can pass an Integer into our Sum class and the compiler handles the translation into an int data value.

We continue to use the base class primarily and not the concrete class of IntegerOf