Sum is an integer
The Sum class represents an integer. We'll refactor this to extend from Integer instead of implementing ISum. This refactor aligns it with the concept of what it is representing.

This also means that we no longer can, or need to, invoke the Value method in our Act section. We have the implicit conversion that will make it an int.