Added IntegerEquality
Our test gets the raw value out so it can compare. This is a behavior. Let's encapsulate that behavior in a new class.

We want to see if something is true or not, which requires a new type, Bool.
Our IntegerEquality class will extend Bool and we will no longer have our test extracting the raw value. The IntegerEquality becomes the only place where the raw values are interacted with. This is how we drive our encapsulation to the extreme. Raw Data is dealt with in very few classes in very limited ways.