Moving classes into the file system
File layout is something that needs to be as heavily refactored as the code itself to stay in sync and continue to make sense.

Here the Interger and Bool files have been moved into a Library folder and type specific folder.

While IntegerEquality is a Bool type, it has a stronger association with, and is returned from a method of, the Integer class. It belongs with the other Interger functionality.