Integers Write
We need to write without exposing our data. To do this we can use a callback.
It might feel hacky, but it maintains encapsulation. There will be big methods
and strange behavior to get and use data from an object. I've had to implement
a hack and it stands out. If you're doing microObjects, that kinda abuse of the
system is going to be obvious.

Our writer has a method and we pass ourself into it. A lot of the time we'll
turn into text first, so we don't need to overload the IWrite interface, but
that will show up later.