Testing the Integer.Into method
The first test was a visual insepect of the test output window to ensure that the value
was written properly. This test has us passing in a fake that we can inspect the value of
to make sure that what we expected to write was actually written. As this is a method
on the abstract class, all Integers get it for free.