Display Data Readme Entry
The Readme was updated with an entry for the DisplayData project. As part of this the
OutputData project has been renamed to Display Data. It's more appropriate.

A future project will be how to send data across the wire, such as an HTTP endpoint
response. This will also extend into storing into a databae. Basically, the techniques of
data transmission to 3rd party components. It'll take a few projects to demonstrate as
each has some uniqueness.