Enhacing SRP of Fluent Type
After adding the Into method onto the Integer. It's clear this is a behavior. It's
going to be less clear to extract this into it's own object so we leave the method as part
of Integer, but need to assign the behavior into it's own interface.  This allows us to
write into anything w/o knowing it's specific type.